Remove support for the compression format `tar.xz`
This format has a very large memory requirement for generation, about 16GB, even with a fallback to a single-threaded compression step, which would be very slow.
This format is not widely used on Windows, and decompression is a pain.
The 7z format has a better balance of size and performance requirements.